Rename file to match export
