+++ lib/extract-props.js
